django-geoaddress/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-geoaddress"
version = "0.1.0"
description = "Django integration for address verification and geocoding. Provides Django fields, widgets, and admin integration for the geoaddress library."
readme = "README.md"
authors = [{ name = "Octolo", email = "dev@octolo.tech" }]
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = [
  "django",
  "geocoding",
  "reverse-geocoding",
  "address",
  "geolocation",
  "maps",
  "django-fields",
  "django-widgets",
  "address-autocomplete",
  "python",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Framework :: Django",
  "Framework :: Django :: 3.2",
  "Framework :: Django :: 4.0",
  "Framework :: Django :: 4.1",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.0",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
  "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
    "Django>=3.2",
    "providerkit>=0.1.0",
    "geoaddress>=0.1.0",
    "django-boosted>=0.1.0",
    "django-providerkit>=0.1.0",
    "django-virtualqueryset>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/hicinformatic/python-geoaddress"
Repository = "https://github.com/hicinformatic/python-geoaddress"
Documentation = "https://github.com/hicinformatic/python-geoaddress/tree/main/django-geoaddress#readme"
Issues = "https://github.com/hicinformatic/python-geoaddress/issues"

[tool.setuptools]
license-files = ["../LICENSE"]

[tool.django-stubs]
django_settings_module = "tests.settings"
